Refactor email_confirmable completely back into the application
Before this we had partially reintegrated the code that had previously
been extracted into email_confirmable. This change does the remaining
work to bring it back into the application. It also does it in a DRY and
Rails Way by using concerns. This should be much more readable and easy
to understand.