Add a spec of questionable value for email confirmable
I just went through the process of setting up a test for this concern
before deciding I didn't need it. However I think we should have tests
for this concern so I've added this as a start.

As I said in the title, I'm not sure how useful this specific one is but
it's a start.