Ensure database tables are UTF-8
This was already done in 3e926d0faca9523d6488525fe110c4866646c6ae but
the server wasn't configured to default to UTF-8 so tables created since
then are the wrong collation.

We've now set the server to default to UTF-8 and this migration will
convert all the tables we've made since the commit mentioned above.

Fixes #1026.