Expand test coverage to corner case
We don't currently have test coverage for the case when the write to
councillor feature is globally enabled, and there are councillors for an
authority, but the feature is turned off on the authority.

This is the case for many councils that have been amalgamated in NSW
recently.

This commit adds test coverage for this situation. It’s currently
failing :(