Stub method to isolate test
Currently these tests are dependent on an environment variable being
set. In my local environment this was set and so tests were passing,
but in Travis CI, they were failing.

This commit stubs the method so these tests don’t need to know about the
environment variable. That configuration is covered by tests on
Authority#write_to_councillors_enabled?.