Fix transient test failure by setting a fixed time
Using Date.today here causes problems when the utc date doesn't
match the local date. This changes the test to use a set Date.new and
freezes the time of the setup to match this. unsubscribed_at has to be
set in the freeze block because the factory generated timestamp runs
outside the freeze.

It's a bit annoying to have all this time stuff in the test for unique
emails, but I can't see a stable way to do it right now.