Make the back amount a round number
While 3.50 is a nice amount to donate, a round number is simpler
conceptually and easier to change. It also means we can use the much
better input modes on mobiles for numbers.

This is also much easier to implement with stripe and make flexible.

Add a todo about adding minimum values and input restrictions to numbers
only to make this as easy as possible.