Remove the PlanningAlerts Pro feature
This feature gave you the ability to enable people getting with three
alerts or more to start a subscription (start paying for them) after a
trial period. If they didn't sign up, their trial would expire and
their alert emails wouldn't have anything applications in them.

We (the OpenAustralia Foundation) didn't find this a very successful way
to make money off the project. It also didn't feel good. We were
particularly going after real estate agents and people using the project
in their work, but we ended up also catching out lots of keen general
citizens.

We're adapting the subscription part of this to enable people who love
the project to start donating to it. This feels much better.