Make the tests more stable by mocking the stripe checkout submission
We don't need to test that checkout.js does what it says, just that the
inputs we control are there for users and submission leads to the
correct processing. These tests currently make an actual call to stripe
which is really slow and difficult to debug here.

This changes the test to have the user put in there details, but we mock
out the actual form submission using a token supplied by stripe-mock.

This is faster and easier to manage.