Improve admin process for confirming a comment for a user
Currently, you control whether a comment is confirmed with a checkbox.
Checking confirm doesn't actually call the confirm method which sets the
confirmed_at time and sends the comment.

This commit replaces the checkbox with a big confirm button on the admin
page for unconfirmed comments and runs the proper action.

Fixes #1069