Fix bug that made initial confirm link 404
When people sign up for the same alert twice (email and address combo),
currently we delete the initial alert and make a new one.

If users then hit the confirm link in the first confirmation email they
received, they get a very confusing 404.

Instead, we should not create a new alert, and just resend the
confirmation email for the initial alert they create. This will mean the
flow will work as expected for users, no matter which email they click
the link in.

Part of https://github.com/openaustralia/planningalerts/issues/117