Update null value in boolean to be false
Currently Alerts can have a confirmed value of nil. This means that what
should be a two value boolean has three values we need to check for.
This means we can't find unconfirmed alerts with queries like:

Alert.where(email: "example", confirmed: false), because there could be
alerts with confirmed = nil.

A boolean should only have two possible values.

This commit adds two migrations. The first sets all alerts with
confirmed=nil to confirmed=false.

The second migration changes the database column to not except a value
of null. The previous commit added a migration to set the column to
default to false. So this shouldn't occur anyway.

Read a little about the tree-value-boolean problem here:
https://robots.thoughtbot.com/avoid-the-threestate-boolean-problem