Only use ssl in production
Currently the ApplicationController is turning on force_ssl in the test
environment, so that we can test that the redirects work.

We then needed to mock the Rails environment across all tests with
javascript so they didn't use ssl.

This commit swaps the exception so that the environment is only changed
for the two tests that check the ssl redirects. In all the tests with
javascript we no longer need to mock the development environment.