Don't unintentionally attempt to re-geocode applications during alerting
Fixes #1117.

When sending alerts we shouldn't expect that applications would attempt
to re-geocode themselves but that has been happening. If an application
had geocoded and got a lat/long but was missing say, a postcode, then
during alert processing it would try to re-geocode.

This is because callbacks on the Application would be run as a
side-effect of incrementing the number of peope alerted. It's arguable
whether applications should be re-geocoded anytime they're saved if
there was just a problem getting their postcode but that's another
issue. We only want to increment this value so we definitely don't want
callbacks to run and that fixes this problem.