Don't use our Google Maps Premium Plan credits for map loads
Working on fixing #1122. We're publicly displaying these maps for free
so there's no reason to use the credits from our Maps Premium Plan. The
Google free plan gives us 25,000 map loads per day so we should be fine
to use this instead.

This also fixes #1098 because it instructs developers to sign up for
their own free key to display maps in development.