Insert application notice below Comments header
This commit aims to fix the issue https://github.com/openaustralia/planningalerts/issues/794
The important information/notice on the application should be present below the Comments header, to catch the attention of the viewer and not be buried at some other place. 
It has been moved to the right position now.