Add TODO
If you're the type of person who recognises what 'JSON' is,
you might well like to sign up for an API key and give us money.

Letting people know that they can do that, in some fairly subtle way
that doesn't impede on the majority of visitors who don't know what JSON
is, could be worthwhile.