Add 'Back PlanningAlerts' banner to bottom of email alerts
This replaces the more subtle call for donations we have now,
and introduces the backers concept.

This is much more obvious visually than previously. We currently get
about 30 people a month clicking through this link currently, I think
we'll get more if we make it more obvious.

This also makes the footer section clearer and simpler without this
donation stuff mixed in.