Fix the test diff by matching how we do the html emails
We then need to add a black line to the bottom of the text email
regression examples, because we're now checking that it exactly matches
what's generated, rather then just that the text is included.
In other words, that extra line was already there, we just weren't
testing for it.