Replace all Date.today calls with Date.current in spec file
This fixes an issue where running the tests in another timezone would
fail.

The fix is to switch from the Ruby to the Rails implementation. The
difference between these two is that Rails uses the configured timezone
of the application to calculate the date. It's recommended not to mix
the Ruby and Rails implementation as things might break:

https://github.com/travisjeffery/timecop#rails-v-ruby-datetime-libraries

We're now consistently using the Rails implementation in this file.