Fix tests that had a problem running in a different timezone
These tests were using Rails' date implementation which uses the
timezone configured in the app (UTC). Whereas the method uses Ruby's
Date.today which uses the system timezone.

We probably expect the latter behaviour so that when the application is
deployed the dates are being compared to the local system date (Sydney),
rather than against UTC. That's why I changed the tests instead of the
method behaviour.

I guess we should be doing something even smarter and checking against
the timezone of where the DA is based because it could be in a different
timezone like WA. But then again most comment periods probably close at
close of business instead of midnight so ¯\_(ツ)_/¯