Remove email from Alert
This removes the email column from the Alert, because it's now available
from the Alert’s alert_subscriber.

This also adds email as an attr_accessor on Alert, so that you can
still initialize an Alert with an "email" attribute, which is used to
assign an AlertSubscriber.

This allows us to maintain our interface for initializing alerts, while
effectively delegating the attribute to another model.