Fix method and test setup
This test setup wasn't doing what we expected, skipping the callback.

I've changed it to manually update the Alerts and remove the
AlertSubscribers so we're certain of what's present.

This test to actually persist the association between the Alert and
AlertSubscriber was actually failing after I fixed the test setup :S