Remove callback
This callback makes the process of creating alerts much less
predictable. Now that we have a validation that the alert must have
an associated alert_subscriber, we can be sure that it will be
associated before the alert is validated, and don't need the callback to
handle this for us.