Don't create AlertSubscriber via alert
Create it before hand. This seems like a simpler way to work, rather
than having to negotiate what data is available in a magical way as the
Alert goes through initialization to save. This way we have to have an
existing AlertSubscriber to make our Alert, and we can depend on it
being there, or the Alert isn't valid.

This commit makes the necessary changes to test setups through out the
app and in the alerts_controller to make this work.