Refactor new contribution process into #new action
Currently the process of building up a new CouncillorContribution is
spread across the #new #create #edit and #update methods of the
CouncillorContributionController. This is overly complex. We can rely on
Rails and the person using the app to replace some of this logic.

This commit changes the "Add another councillor" button to send the user
back to the #new action, rather than the #create action.

To enable us to post data to the #new action, we've added a new route.

The existing councillors are passed to the view as hidden fields so they
can be passed on with the submit request.

This enables us to remove the #edit and #update actions from the
controller and the associated routes. We've also removed the unused
edit.html.haml view file.

We've also removed the logic for distinguishing between 'add councillor'
requests and submit requests from the #create action, because it now
only receives submit requests to create the contribution.