Split integration spec file in two
This creates a new file that covers the different options for
contribution for providing their information to us.

We split this out to make the original spec file more focused and easier
to understand.

This now means there isn't a single spec for working through the whole
contribution process from beginning to end. But we think that's ok for now,
because the contribution and the stage of providing details are quite separate.