Change validation to make classes more modular
Requiring the CouncillorContribution to have at least one associated
SuggestCouncillor made test setup very complex and binds these two
classes too tightly. You couldn't create a SuggestedCouncillor without
having a CouncillorContribution built already to associate, and the same
in reverse. This is causing spec/features/admin_views_councillor_contribution_spec.rb
to break because we weren't set up to handle this double dependency in the
factories and it's complicated to implement.

I think the purpose of the validation was to prevent someone from
submitting a single invalid councillor as there CouncillorContribution.

This commit changes the validation on CouncillorCountribution to make sure
validations are run on any associated suggested councillors.
This should prevent people from creating a contribution with no councillors,
because there's no option to submit no councillors in our form.

I've added a test for this new validation that uses mocking to isolate
the test from any changes to the validations we might make on SuggestedCouncillor.

I've also added a factory we can use for CouncillorContributions as part of this change.