Confirm that the original councillor wasn't saved
This is a bit hacky, it's best to just test what you see in integration
tests (see http://www.betterspecs.org/#integration).

But I think in this case we need to make it clear in the test what was
actually saved to the database, because the user doesn't see a
confirmation we can use to confirm that the correct councillor was
submitted and the original one that was edited wasn't.