Changed the order of factory datebase setup
Currently test in faiing on Travis C.I., for dupricated definition eror,
although there is no duprication in the file. Also it is giving a error for
line that is blank. I am changing the order of the factory to see if the
error for the line number would change.