Make the instructions collapsed by default
This allows people to call them up as needed and hide them away when
they don't.

This uses a progressive enhancement approach using the <details>
element. Browsers that don't recognise that element will just show the
instructions and the user wont have the option to hide them.

According to https://caniuse.com/#search=details 89% of browsers in
active use in Australia support the element.