Inline method
I don't think this extraction is much of an improvement on the long
string construction, or not one that warrants adding a method to the
class.

The more obvious refactor here, for later, is to extract the full string
into a method on CouncillorContribution like #title or something like that.