Invite contributors to help improve the instructions
I think this is an important element of our principle to "Share ownership"
with the people contributing. They're likely to become more knowledgable
about the contribution process than we are, so we should invite them to
share that knowledge.

See Principles: https://github.com/openaustralia/australian_local_councillors_popolo/#design-principles

This is just in a small way, but a nice touch I think.