Move validation message tests to view spec
Having these tests to make sure validation messages are shown clutters
the integration tests I think.

We're still checking in the integration tests that the person doesn't
progress to the next page, but I don't think we need to tests for all
the kinds of validations there. That feels too much like we're testing
the model validations, where as we have those tests and here we just
want to make sure that the error messages are rendered.

I've included a check against both fields to make sure they're both
rendered.