Fail fast, blow up instead of showing error to user
We don't know of a good reason why this contributor would not be able to
be saved, so raise an error here.

This is actually what would already happen here because we're using
`create_contributor!` not `create_contributor`. This will raise an error
on failure rather than returning false.

Removing this code makes that clearer.