Fail fast, don't return message if contribution can't update
Given that there are no validations on the source attribute, there's no
obvious reason why the contribution would not be able to update here.

We should actually fail here and raise an error so admins are notified
via honeybadger rather than asking the user to try again—if things
aren't working that's probably not going to help them.