Make time testing more deterministic
Replace the need for Timecop by setting the UTC create time and testing
that it's present, rather than freezing time.

In the current situation, the tests would pass if the admin page showed
the current time, rather than the time the contribution was created,
because they're the same and we can't tell the difference.