Add failing test for marking contribution accepted
Marking contributions as 'reviewed' doesn't provide enough information
to admins or people who want to consume the contribution data from the API.

People consuming the councillor contribution need to know not only that
someone reviewed the contribution, but that they found it to be
acceptable. As a consumer, I don't want to pull over unacceptable
contributions. Currently I can't tell if they are.