Seperate web/email API keys
Google's API dashboard lets us break down usage by API key, but not by
anything else that helps us track where the usage is coming from.

The intent here is to allow us to differentiate between straight web
traffic (which will use the new key) and links in alert emails (which
will continue to use the old key).

Why this way around? Because lots of emails have already been sent out
with the old key. We can't change those easily; but we *can* easily
change what we display on the website. Once this change is live we can
be fairly confident that all web traffic will switch to the new key
immediately, while anything using the old credential will be links
from emails.

This change requires a matching change in the infrastructure
repo (although due to the default values provided here, this change
should be fine to push without the matching change in infrastructure).