Remove link to preview page
For the time being we decided that having a preview for the emails that have gone out is a surprising amount of work to do accurately (tracking versions of emails, etc..) So better to focus on the preview as part of the comment writing process instead. We can always can come back to this later.